Removed custom auth user model, extended default
