--- ris/ris/models/__init__.py
@@ -1,1 +1,0 @@
-from models import Person, User, FamilyDoctor, PacsImage, RadiologyRecord
